# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FlowVerse 3D HTTP API
  version: 0.1.0
  summary: API para diseño, versionado, validación y simulación de flujos.
  description: |
    Contrato HTTP canónico del MVP. La autenticación utiliza cookies HttpOnly
    opacas. Toda mutación autenticada debe incluir `X-CSRF-Token`. Los errores
    comparten un sobre estable y cada respuesta incluye `X-Request-Id`.
servers:
  - url: http://localhost:8080
    description: Desarrollo local
...
security:
  - cookieAuth: []
paths:
...
  /v1/flow-versions/{versionId}/runs:
    parameters:
      - $ref: "#/components/parameters/VersionId"
    post:
      tags: [Runs]
      operationId: createRun
      summary: Inicia una simulación determinista.
      description: |
        Crea una ejecución sobre una versión publicada e inmutable. La clave
        idempotente tiene alcance por usuario autenticado, versionId y key
        durante 24 horas. Repetir el mismo body devuelve el run existente con
        200; reutilizar la key en ese alcance con otro body devuelve
        `409 idempotency.payload_mismatch`.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/simulation-request.schema.json"
      responses:
        "200":
          description: Replay idempotente; no se crea ni inicia otro run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "201":
          description: Ejecución creada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /v1/flows/{flowId}/runs:
    parameters:
      - $ref: "#/components/parameters/FlowId"
    post:
      tags: [Runs]
      operationId: createDraftRun
      summary: Simula un snapshot exacto del borrador actual.
      description: |
        Captura el documento y ETag actuales antes de simular; cambios
        posteriores no alteran el run. Esta extensión no exige `If-Match`.
        La clave idempotente tiene alcance por usuario autenticado, flowId,
        ETag capturado y key durante 24 horas. El mismo body dentro del alcance
        devuelve 200; otro body devuelve
        `409 idempotency.payload_mismatch`. Una nueva revisión del borrador
        constituye un target distinto y puede reutilizar la misma key.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/simulation-request.schema.json"
      responses:
        "200":
          description: Replay idempotente; no se crea ni inicia otro run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "201":
          description: Run creado a partir del snapshot del borrador.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /v1/runs/{runId}:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      tags: [Runs]
      operationId: getRun
      summary: Obtiene estado y resumen de una ejecución.
      responses:
        "200":
          description: Ejecución.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/runs/{runId}/pause:
    parameters:
      - $ref: "#/components/parameters/RunId"
    post:
      tags: [Runs]
      operationId: pauseRun
      summary: Pausa la ejecución al terminar la transición actual.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "202":
          $ref: "#/components/responses/RunAccepted"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/runs/{runId}/resume:
    parameters:
      - $ref: "#/components/parameters/RunId"
    post:
      tags: [Runs]
      operationId: resumeRun
      summary: Reanuda una ejecución pausada.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "202":
          $ref: "#/components/responses/RunAccepted"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/runs/{runId}/step:
    parameters:
      - $ref: "#/components/parameters/RunId"
    post:
      tags: [Runs]
      operationId: stepRun
      summary: Completa una visita de nodo y vuelve a pausa.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "202":
          $ref: "#/components/responses/RunAccepted"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/runs/{runId}/cancel:
    parameters:
      - $ref: "#/components/parameters/RunId"
    post:
      tags: [Runs]
      operationId: cancelRun
      summary: Cancela una ejecución activa.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "202":
          $ref: "#/components/responses/RunAccepted"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /v1/flows/{flowId}/share-links:
    parameters:
      - $ref: "#/components/parameters/FlowId"
    get:
      tags: [Shares]
      operationId: listShareLinks
      summary: Lista enlaces públicos sin exponer sus tokens.
      responses:
        "200":
          description: Enlaces.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShareLink"
    post:
      tags: [Shares]
      operationId: createShareLink
      summary: Comparte una versión y, opcionalmente, runs completados.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShareLinkRequest"
      responses:
        "201":
          description: El token se entrega una sola vez.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedShareLink"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: flowverse_access
...
  parameters:
...
    FlowId:
      name: flowId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VersionId:
      name: versionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Clave obligatoria, recortada antes de validarse y de 8 a 128
        caracteres. Se conserva 24 horas. Su alcance combina el usuario
        autenticado, el target exacto de la operación y esta key.
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/Issue"
    Issue:
      type: object
      additionalProperties: false
      required: [code, severity, message]
      properties:
        code:
          type: string
        severity:
          enum: [info, warning, error]
        message:
          type: string
        path:
          type: string
        nodeId:
          type: string
        edgeId:
          type: string
...
    Run:
      type: object
      additionalProperties: false
      required:
        [id, flowVersionId, status, triggerNodeId, logicalTimeMs, createdAt]
      properties:
        id:
          type: string
          format: uuid
        traceId:
          type: string
          pattern: "^[0-9a-f]{32}$"
        flowVersionId:
          description: UUID de la versión publicada; vacío para un snapshot de borrador.
          oneOf:
            - type: string
              format: uuid
            - type: string
              const: ""
        status:
          enum:
            [queued, running, paused, completed, failed, cancelled, interrupted]
        triggerNodeId:
          type: string
        logicalTimeMs:
          type: integer
          minimum: 0
        playbackSpeed:
          type: number
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: [string, "null"]
          format: date-time
        finishedAt:
          type: [string, "null"]
          format: date-time
        result:
          type: [object, "null"]
...
    CreateShareLinkRequest:
      type: object
      additionalProperties: false
      required: [versionId, runIds]
      properties:
        versionId:
          type: string
          format: uuid
        runIds:
          type: array
          maxItems: 20
          uniqueItems: true
          items:
            type: string
            format: uuid
        expiresAt:
          type: [string, "null"]
          format: date-time
    ShareLink:
      type: object
      additionalProperties: false
      required: [id, flowId, versionId, runIds, createdAt, revoked]
      properties:
        id:
          type: string
          format: uuid
        flowId:
          type: string
          format: uuid
        versionId:
          type: string
          format: uuid
        runIds:
          type: array
          items:
            type: string
            format: uuid
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: [string, "null"]
          format: date-time
        revoked:
          type: boolean
    CreatedShareLink:
      type: object
      additionalProperties: false
      required:
        [id, flowId, versionId, runIds, createdAt, revoked, token, publicUrl]
      properties:
        id:
          type: string
          format: uuid
        flowId:
          type: string
          format: uuid
        versionId:
          type: string
          format: uuid
        runIds:
          type: array
          items:
            type: string
            format: uuid
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: [string, "null"]
          format: date-time
        revoked:
          type: boolean
        token:
          type: string
        publicUrl:
          type: string
          format: uri
...
  responses:
...
    RunAccepted:
      description: La orden de control fue aceptada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Run"
    BadRequest:
      description: |
        Solicitud inválida. En creación de runs incluye
        `idempotency.key_required`, `idempotency.key_invalid`, JSON inválido u
        overrides inexistentes.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Sesión inexistente, vencida o revocada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: El rol no permite la operación.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Recurso inexistente o no visible para el usuario.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Estado actual incompatible con la operación.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    IdempotencyConflict:
      description: |
        La misma Idempotency-Key ya fue usada dentro de su alcance con otro
        body. El error estable es `idempotency.payload_mismatch`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationFailed:
      description: Entrada o grafo inválido.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    PayloadTooLarge:
      description: El documento supera 24 MiB o los límites del grafo.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...